Extract behavior as a command.
