+++ src/main/java/oripa/domain/paint/outline/SelectingVertexForOutline.java
@@ -1,20 +1,12 @@
 package oripa.domain.paint.outline;
 
-import java.util.Collection;
-
 import javax.vecmath.Vector2d;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import oripa.domain.cptool.Painter;
 import oripa.domain.paint.PaintContext;
 import oripa.domain.paint.core.PickingVertex;
-import oripa.geom.GeomUtil;
+import oripa.util.Command;
 
 public class SelectingVertexForOutline extends PickingVertex {
-	private static final Logger logger = LoggerFactory.getLogger(SelectingVertexForOutline.class);
-
 	private final CloseTempOutlineFactory closeTempOutlineFactory;
 
 	/**
@@ -39,32 +31,7 @@ public class SelectingVertexForOutline extends PickingVertex {
 
 	@Override
 	protected void onResult(final PaintContext context, final boolean doSpecial) {
-
-		logger.debug("# of picked vertices (before): " + context.getPickedVertices().size());
-
-		Vector2d v = context.popVertex();
-
-		if (context.getPickedVertices().stream()
-				.anyMatch(tv -> GeomUtil.distance(v, tv) < 1)) {
-			if (context.getVertexCount() > 2) {
-				// finish editing
-
-				context.creasePatternUndo().pushUndoInfo();
-				closeTmpOutline(context.getPickedVertices(), context.getPainter());
-
-				context.clear(false);
-				context.setMissionCompleted(true);
-			}
-		} else {
-			// continue selecting
-			context.pushVertex(v);
-		}
-
-		logger.debug("# of picked vertices (after): " + context.getPickedVertices().size());
+		Command command = new OutlineEditerCommand(context, closeTempOutlineFactory);
+		command.execute();
 	}
-
-	private void closeTmpOutline(final Collection<Vector2d> outlineVertices, final Painter painter) {
-		closeTempOutlineFactory.create().execute(outlineVertices, painter);
-	}
-
 }
